requirements.txt
# Core Dependencies
python-dotenv==1.1.1
gradio>=4.0.0

# LLM & API
google-genai>=1.0.0
google-generativeai>=0.8.0
youtube-transcript-api==1.2.3

# Audio/ASR
faster-whisper>=1.0.0
yt-dlp>=2024.1.0
torch>=2.0.0

# Data Processing & Validation
pydantic==2.11.9
pydantic-core==2.33.2
numpy==1.26.4
pandas==2.2.3

# HTTP & Requests
requests==2.32.5
aiohttp==3.12.14

# Development & Testing
jupyter==1.0.0
ipython==9.0.2
pytest==7.4.4
black==24.1.1
flake8==7.0.0
mypy==1.7.1

# Utilities
python-dateutil==2.9.0.post0
